adding create-server and create-app to simplify the code structure
