+++ lib/acl.js
@@ -9,7 +9,7 @@ var S = require('string');
 var url = require('url');
 var async = require('async');
 
-var debug = require('./logging').ACL;
+var debug = require('./debug').ACL;
 var utils = require('./utils.js');
 var ns = require('./vocab/ns.js').ns;
 var rdfVocab = require('./vocab/rdf.js');
