+++ lib/handlers/delete.js
@@ -2,7 +2,7 @@
 "use strict";
 
 var fs = require('fs');
-var debug = require('../logging').handlers;
+var debug = require('../debug').handlers;
 var utils = require('../utils.js');
 var metadata = require('../metadata.js');
