+++ lib/handlers/get.js
@@ -9,7 +9,7 @@ var $rdf = require('rdflib');
 var S = require('string');
 var async = require('async');
 
-var debug = require('../logging').handlers;
+var debug = require('../debug').handlers;
 var acl = require('../acl.js');
 var header = require('../header.js');
 var metadata = require('../metadata.js');
