+++ lib/handlers/patch.js
@@ -5,7 +5,7 @@ var mime = require('mime');
 var fs = require('fs');
 var $rdf = require('rdflib');
 
-var debug = require('../logging').handlers;
+var debug = require('../debug').handlers;
 var utils = require('../utils.js');
 var subscription = require('../subscription.js');
 var HttpError = require('../http-error');
