+++ lib/handlers/post.js
@@ -5,7 +5,7 @@ var path = require('path');
 var $rdf = require('rdflib');
 var S = require('string');
 
-var debug = require('../logging').handlers;
+var debug = require('../debug').handlers;
 var utils = require('../utils.js');
 var header = require('../header.js');
 var metadata = require('../metadata.js');
