+++ lib/handlers/put.js
@@ -5,7 +5,7 @@ var mime = require('mime');
 var path = require('path');
 var $rdf = require('rdflib');
 
-var debug = require('../logging').handlers;
+var debug = require('../debug').handlers;
 var utils = require('../utils.js');
 var header = require('../header.js');
