+++ lib/ldp.js
@@ -11,7 +11,7 @@ var fs = require('fs');
 var mkdirp = require('fs-extra').mkdirp;
 var uuid = require('node-uuid');
 
-var debug = require('./logging');
+var debug = require('./debug');
 
 var utils = require('./utils.js');
 var ns = require('./vocab/ns.js').ns;
@@ -36,13 +36,6 @@ function LDP(argv) {
   ldp.port = argv.port;
   ldp.secret = argv.secret;
   ldp.cert = argv.cert;
-  ldp.key = argv.key;
-  ldp.mount = argv.mount || '/';
-  // Removing ending '/'
-  if (ldp.mount.length > 1 &&
-    ldp.mount[ldp.mount.length - 1] === '/') {
-    ldp.mount = ldp.mount.slice(0, -1);
-  }
 
   ldp.verbose = argv.verbose;
   ldp.webid = argv.webid;
@@ -54,7 +47,6 @@ function LDP(argv) {
   ldp.suffixChanges = argv.suffixChanges || '.changes';
   ldp.suffixSSE = argv.suffixSSE || '.events';
 
-  ldp.pathFilter = regexp().start(ldp.mount).toRegExp();
   ldp.proxy = argv.proxy;
 
   // Cache of usedURIs
@@ -80,7 +72,6 @@ function LDP(argv) {
     ldp.skin = argv.skin || 'https://linkeddata.github.io/warp/#/list/';
   }
 
-  debug.settings("mount: " + ldp.mount);
   debug.settings("root: " + ldp.root);
   debug.settings("URI path filter regexp: " + ldp.pathFilter);
   debug.settings("Verbose: " + !!ldp.verbose);
