+++ lib/login.js
@@ -2,7 +2,7 @@
 "use strict";
 
 var webid = require('webid');
-var debug = require('./logging').login;
+var debug = require('./debug').login;
 var HttpError = require('./http-error');
 
 function loginHandler(req, res, next) {
