+++ lib/parse.js
@@ -4,7 +4,7 @@
 var N3 = require('n3');
 var jsonld = require('jsonld');
 var async = require('async');
-var debug = require('./logging').parse;
+var debug = require('./debug').parse;
 
 function parseHandler(req, res, next) {
     convertToTurtle(req.text, req, function(err, result) {
