+++ lib/subscription.js
@@ -3,7 +3,7 @@
 
 var $rdf = require('rdflib');
 var redis = require('redis');
-var debug = require('./logging').subscription;
+var debug = require('./debug').subscription;
 var utils = require('./utils.js');
 var PATCH = $rdf.Namespace('http://www.w3.org/ns/pim/patch#');
